run() assembly 3: R11-R13 on the shared state; committed usage as uint16
Committed usage is uint16 with the reference charge arithmetic (sum in double, truncating). TreeRoute
carries the maze route. maze_phase.rs: convertToMazeroute, routeMonotonic/All, the LV rounds (exposing
each round), and the pre-loop resets. The ripup gate and monotonic router split into a pure decision
and a write. Chain vs the boundaries through R13: 138/146 runs exact; congestion3/4 stop at the usage
check as the reference does. Constructed cases for the cost-table height and the LV schedule.

+++ tests/monotonic.rs
@@ -140,7 +140,7 @@ fn midpoint_and_orientations_match_the_reference() {
         let mut charged: Vec<(i32, i32, f64, bool)> = Vec::new();
         for y in 0..grid.y_grids {
             for x in 0..grid.h_columns() {
-                let u = grid.usage_h(x, y);
+                let u = f64::from(grid.usage_h(x, y));
                 if u != 0.0 {
                     charged.push((x as i32, y as i32, u, true));
                 }
@@ -148,7 +148,7 @@ fn midpoint_and_orientations_match_the_reference() {
         }
         for y in 0..grid.v_rows() {
             for x in 0..grid.x_grids {
-                let u = grid.usage_v(x, y);
+                let u = f64::from(grid.usage_v(x, y));
                 if u != 0.0 {
                     charged.push((x as i32, y as i32, u, false));
                 }
